#2125. Gather helm repo related functions behind a new interface RepositoryStore
