+++ internal/federation/deployment.go
@@ -85,7 +85,7 @@ func InstallOrUpgradeDeployment(
 			if !upgrade {
 				return nil
 			}
-			_, err = helm.UpgradeDeployment(releaseName, deploymentName, chartVersion, nil, values, false, kubeConfig, helm.GenerateHelmRepoEnv(org.Name))
+			_, err = helm.UpgradeDeployment(org.Name, releaseName, deploymentName, chartVersion, nil, values, false, kubeConfig)
 			if err != nil {
 				return emperror.WrapWith(err, "could not upgrade deployment", "deploymentName", deploymentName)
 			}
@@ -104,6 +104,7 @@ func InstallOrUpgradeDeployment(
 	}
 
 	_, err = helm.CreateDeployment(
+		org.Name,
 		deploymentName,
 		chartVersion,
 		nil,
@@ -112,7 +113,6 @@ func InstallOrUpgradeDeployment(
 		false,
 		nil,
 		kubeConfig,
-		helm.GenerateHelmRepoEnv(org.Name),
 		options...,
 	)
 	if err != nil {
